+++ CBM_Benchmark/Main.cpp
@@ -86,11 +86,11 @@ main(int argc, char *argv[])
         }
         printf("Endpoints: %d\n", config->num_ep);
 
-        int total_data_size = atoi(argv[3]);
+        double total_data_size = atof(argv[3]);
         if (total_data_size > 0) {
             config->total_data_size = total_data_size * 1024 * 1024;
         }
-        printf("total_data_size: %d\n", config->total_data_size);
+        printf("total_data_size: %f\n", config->total_data_size);
 
         char *addr = argv[argc-1];
         addr = host2ip::resolve(addr);
